+++ subscription/util/blocking/competing-consumer-subscription/src/test/java/org/occurrent/subscription/blocking/competingconsumers/CompetingConsumerSubscriptionModelChaosTest.java
@@ -107,7 +107,7 @@ class CompetingConsumerSubscriptionModelChaosTest {
         Retry retry = RetryStrategy
                 .fixed(1000)
                 .retryIf(not(DuplicateCloudEventException.class::isInstance))
-                .onError(((retryInfo, throwable) -> log.warn("Detected error (retryCount={}). Error={} - {}", retryInfo.getRetryCount(), throwable.getClass().getSimpleName(), throwable.getMessage())));
+                .onBeforeRetry(((retryInfo, throwable) -> log.warn("Detected error (retryCount={}). Error={} - {}", retryInfo.getRetryCount(), throwable.getClass().getSimpleName(), throwable.getMessage())));
 
         Thread eventPublishingThread = new Thread(() -> {
             // When